Update Bookitem file
