+++ src/features/Books/BookItem.jsx
@@ -1,7 +1,7 @@
-import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
+import React, { useState, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
-import { AddBook } from '../../redux/books/books';
+import { AddBook, LoadBooks, RemoveBook } from '../../redux/books/books';
 
 const BookItem = () => {
   const valueInitialState = {
@@ -13,7 +13,17 @@ const BookItem = () => {
   const [values, setValues] = useState(valueInitialState);
   const [submitted, setSubmitted] = useState(false);
 
+  const books = useSelector((state) => state.book.books[0]);
   const dispatch = useDispatch();
+
+  const handleRemove = (id) => {
+    dispatch(RemoveBook(id));
+    dispatch(LoadBooks());
+  };
+
+  useEffect(() => {
+    dispatch(LoadBooks());
+  }, [values, submitted]);
 
   const handleChange = (e) => {
     const { name, value } = e.target;
@@ -35,6 +45,7 @@ const BookItem = () => {
         category: category || 'Action',
       };
       dispatch(AddBook(bookArray));
+      dispatch(LoadBooks());
       setSubmitted(true);
       setValues('');
     }
@@ -42,13 +53,34 @@ const BookItem = () => {
 
   return (
     <div>
+      {books
+          && books.map((book) => (
+            <li key={book.item_id}>
+              {book.title}
+              <br />
+              {book.author}
+              <br />
+              {book.category}
+              <br />
+              <button type="button">Comments</button>
+              <button
+                type="button"
+                onClick={() => handleRemove(book.item_id)}
+              >
+                Remove
+              </button>
+              <button type="button">Edit</button>
+            </li>
+          ))}
+      <h3 className="text"> Add Book Name! </h3>
+
       {submitted ? (
         <div>
           <h4>You submitted successfully!</h4>
           <button
             type="submit"
             className="btn btn-success"
-            onClick={handleSubmitAnother}
+            onClick={() => handleSubmitAnother()}
           >
             Add Another
           </button>
@@ -74,7 +106,6 @@ const BookItem = () => {
 
           <label htmlFor="categoryId" className="form-label">
             Book Category
-            {' '}
             {' '}
             <select
               name="category"
